PR-04a: extract args.rs and autostart.rs
Behaviour-preserving relocation:
- args.rs (new): TrayFocusOnly, Args (clap derive), parse_dbus_suffix_arg,
  resolve_install_gnome_extension, resolve_control_command.
- autostart.rs (new): the 8 desktop-entry helpers plus the 3 AUTOSTART_*
  constants.

main.rs: 7092 -> 6788 LOC.

Verification: cargo build clean; cargo test --bin kanata-switcher
261 passed / 0 failed.
