+++ src/daemon/main.rs
@@ -93,6 +93,8 @@ mod environ;
 mod dbus_naming;
 mod config;
 mod focus;
+mod args;
+mod autostart;
 
 use constants::*;
 use errors::DynError;
@@ -100,10 +102,12 @@ use environ::*;
 use dbus_naming::*;
 use config::*;
 use focus::*;
+use args::*;
+use autostart::*;
 
 #[cfg(test)]
 #[allow(unused_imports)]
-pub(crate) use crate::{constants::*, errors::*, environ::*, dbus_naming::*, config::*, focus::*};
+pub(crate) use crate::{constants::*, errors::*, environ::*, dbus_naming::*, config::*, focus::*, args::*, autostart::*};
 
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -129,309 +133,6 @@ impl ControlCommand {
             ControlCommand::Unpause => "unpause",
         }
     }
-}
-
-// === CLI ===
-
-#[derive(Clone, Copy, Debug, ValueEnum)]
-enum TrayFocusOnly {
-    True,
-    False,
-}
-
-impl TrayFocusOnly {
-    fn as_bool(self) -> bool {
-        matches!(self, TrayFocusOnly::True)
-    }
-
-    fn as_arg(self) -> &'static str {
-        match self {
-            TrayFocusOnly::True => "true",
-            TrayFocusOnly::False => "false",
-        }
-    }
-}
-
-#[derive(Parser)]
-#[command(name = "kanata-switcher")]
-#[command(about = "Switch kanata layers based on focused window")]
-struct Args {
-    #[arg(short = 'p', long, default_value = "10000")]
-    port: u16,
-
-    #[arg(short = 'H', long, default_value = "127.0.0.1")]
-    host: String,
-
-    #[arg(short = 'c', long)]
-    config: Option<PathBuf>,
-
-    /// Quiet mode: suppress focus and layer-switch messages
-    #[arg(short = 'q', long)]
-    quiet: bool,
-
-    /// Suppress focus messages only
-    #[arg(long)]
-    quiet_focus: bool,
-
-    /// Auto-install GNOME extension if missing (default behavior)
-    #[arg(long)]
-    install_gnome_extension: bool,
-
-    /// Do not auto-install GNOME extension
-    #[arg(long)]
-    no_install_gnome_extension: bool,
-
-    /// Disable the StatusNotifier (SNI) indicator on non-GNOME desktops
-    #[arg(long)]
-    no_indicator: bool,
-
-    /// Override SNI focus-only mode (true/false). When set, GSettings is not read.
-    #[arg(long, value_enum, value_name = "true|false")]
-    indicator_focus_only: Option<TrayFocusOnly>,
-
-    /// Install autostart desktop entry and exit
-    #[arg(long, conflicts_with_all = ["uninstall_autostart", "restart", "pause", "unpause"])]
-    install_autostart: bool,
-
-    /// Uninstall autostart desktop entry and exit
-    #[arg(long, conflicts_with_all = ["install_autostart", "restart", "pause", "unpause"])]
-    uninstall_autostart: bool,
-
-    /// Send Restart request to an existing daemon and exit
-    #[arg(long, conflicts_with_all = ["pause", "unpause"])]
-    restart: bool,
-
-    /// Send Pause request to an existing daemon and exit
-    #[arg(long, conflicts_with_all = ["restart", "unpause"])]
-    pause: bool,
-
-    /// Send Unpause request to an existing daemon and exit
-    #[arg(long, conflicts_with_all = ["restart", "pause"])]
-    unpause: bool,
-
-    /// DBus suffix used as the last name element of this daemon's well-known bus
-    /// name (`com.github.kanata.Switcher.instances.<suffix>`). When omitted the
-    /// suffix is auto-derived from `--host` and `--port`. Control CLI commands
-    /// target this exact suffix when set; otherwise they broadcast.
-    #[arg(long, value_name = "SUFFIX", value_parser = parse_dbus_suffix_arg)]
-    dbus_suffix: Option<String>,
-}
-
-fn parse_dbus_suffix_arg(raw: &str) -> Result<String, String> {
-    sanitize_dbus_suffix(raw).map_err(|error| error.to_string())
-}
-
-const AUTOSTART_DESKTOP_FILENAME: &str = "kanata-switcher.desktop";
-const AUTOSTART_PASSTHROUGH_OPTIONS: &[&str] = &[
-    "port",
-    "host",
-    "config",
-    "quiet",
-    "quiet_focus",
-    "install_gnome_extension",
-    "no_install_gnome_extension",
-    "no_indicator",
-    "indicator_focus_only",
-    "dbus_suffix",
-];
-const AUTOSTART_ONESHOT_OPTIONS: &[&str] = &[
-    "restart",
-    "pause",
-    "unpause",
-    "install_autostart",
-    "uninstall_autostart",
-];
-
-fn resolve_install_gnome_extension(matches: &ArgMatches) -> bool {
-    use clap::parser::ValueSource;
-
-    let install_from_cli =
-        matches.value_source("install_gnome_extension") == Some(ValueSource::CommandLine);
-    let no_install_from_cli =
-        matches.value_source("no_install_gnome_extension") == Some(ValueSource::CommandLine);
-
-    match (install_from_cli, no_install_from_cli) {
-        (false, false) => true,
-        (true, false) => true,
-        (false, true) => false,
-        (true, true) => {
-            let install_idx = matches.index_of("install_gnome_extension");
-            let no_install_idx = matches.index_of("no_install_gnome_extension");
-            match (install_idx, no_install_idx) {
-                (Some(i), Some(n)) => i > n,
-                _ => true,
-            }
-        }
-    }
-}
-
-fn resolve_control_command(args: &Args) -> Option<ControlCommand> {
-    if args.restart {
-        return Some(ControlCommand::Restart);
-    }
-    if args.pause {
-        return Some(ControlCommand::Pause);
-    }
-    if args.unpause {
-        return Some(ControlCommand::Unpause);
-    }
-    None
-}
-
-fn resolve_binary_path() -> Result<PathBuf, Box<dyn std::error::Error + Send + Sync>> {
-    let exe_path = env::current_exe()?;
-    let canonical = exe_path.canonicalize()?;
-    Ok(canonical)
-}
-
-fn autostart_dir() -> Result<PathBuf, Box<dyn std::error::Error + Send + Sync>> {
-    if let Ok(xdg_config_home) = env::var("XDG_CONFIG_HOME") {
-        if xdg_config_home.is_empty() {
-            return Err("XDG_CONFIG_HOME is empty".into());
-        }
-        return Ok(PathBuf::from(xdg_config_home).join("autostart"));
-    }
-    let home = env::var("HOME")?;
-    if home.is_empty() {
-        return Err("HOME is empty".into());
-    }
-    Ok(PathBuf::from(home).join(".config").join("autostart"))
-}
-
-fn autostart_desktop_path() -> Result<PathBuf, Box<dyn std::error::Error + Send + Sync>> {
-    Ok(autostart_dir()?.join(AUTOSTART_DESKTOP_FILENAME))
-}
-
-fn escape_desktop_exec_arg(value: &str) -> String {
-    let mut escaped = String::with_capacity(value.len() + 2);
-    escaped.push('"');
-    for ch in value.chars() {
-        match ch {
-            '\\' => escaped.push_str("\\\\"),
-            '"' => escaped.push_str("\\\""),
-            '%' => escaped.push_str("%%"),
-            '\n' => escaped.push_str("\\n"),
-            '\r' => escaped.push_str("\\r"),
-            '\t' => escaped.push_str("\\t"),
-            _ => escaped.push(ch),
-        }
-    }
-    escaped.push('"');
-    escaped
-}
-
-fn build_autostart_desktop_content(exec_path: &Path, exec_args: &[String]) -> String {
-    let exec_path_str = exec_path
-        .to_str()
-        .expect("autostart exec path contains invalid UTF-8");
-    let mut exec_parts = Vec::with_capacity(exec_args.len() + 1);
-    exec_parts.push(escape_desktop_exec_arg(exec_path_str));
-    for arg in exec_args {
-        exec_parts.push(escape_desktop_exec_arg(arg));
-    }
-    let exec_line = exec_parts.join(" ");
-    format!(
-        "[Desktop Entry]\nType=Application\nName=Kanata Switcher\nExec={}\nTryExec={}\nX-GNOME-Autostart-enabled=true\n",
-        exec_line,
-        escape_desktop_exec_arg(exec_path_str)
-    )
-}
-
-fn autostart_passthrough_args(matches: &ArgMatches, args: &Args) -> Vec<String> {
-    use clap::parser::ValueSource;
-
-    let mut exec_args = Vec::new();
-
-    for &name in AUTOSTART_PASSTHROUGH_OPTIONS {
-        if matches.value_source(name) != Some(ValueSource::CommandLine) {
-            continue;
-        }
-        match name {
-            "port" => {
-                exec_args.push("-p".to_string());
-                exec_args.push(args.port.to_string());
-            }
-            "host" => {
-                exec_args.push("-H".to_string());
-                exec_args.push(args.host.clone());
-            }
-            "config" => {
-                let config = args
-                    .config
-                    .as_ref()
-                    .expect("config missing after command-line input");
-                exec_args.push("-c".to_string());
-                exec_args.push(config.to_string_lossy().to_string());
-            }
-            "quiet" => {
-                exec_args.push("-q".to_string());
-            }
-            "quiet_focus" => {
-                exec_args.push("--quiet-focus".to_string());
-            }
-            "install_gnome_extension" => {
-                exec_args.push("--install-gnome-extension".to_string());
-            }
-            "no_install_gnome_extension" => {
-                exec_args.push("--no-install-gnome-extension".to_string());
-            }
-            "no_indicator" => {
-                exec_args.push("--no-indicator".to_string());
-            }
-            "indicator_focus_only" => {
-                let value = args
-                    .indicator_focus_only
-                    .expect("indicator_focus_only missing after command-line input");
-                exec_args.push("--indicator-focus-only".to_string());
-                exec_args.push(value.as_arg().to_string());
-            }
-            "dbus_suffix" => {
-                let suffix = args
-                    .dbus_suffix
-                    .as_ref()
-                    .expect("dbus_suffix missing after command-line input");
-                exec_args.push("--dbus-suffix".to_string());
-                exec_args.push(suffix.clone());
-            }
-            _ => {
-                panic!("autostart passthrough option missing handler: {}", name);
-            }
-        }
-    }
-
-    exec_args
-}
-
-fn install_autostart_desktop(
-    matches: &ArgMatches,
-    args: &Args,
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    for option in AUTOSTART_ONESHOT_OPTIONS {
-        if AUTOSTART_PASSTHROUGH_OPTIONS.contains(option) {
-            return Err(format!("autostart option lists overlap: {}", option).into());
-        }
-    }
-    let exec_path = resolve_binary_path()?;
-    let exec_args = autostart_passthrough_args(matches, args);
-    let content = build_autostart_desktop_content(&exec_path, &exec_args);
-
-    let autostart_dir = autostart_dir()?;
-    std::fs::create_dir_all(&autostart_dir)?;
-    let desktop_path = autostart_dir.join(AUTOSTART_DESKTOP_FILENAME);
-
-    std::fs::write(&desktop_path, content)?;
-    println!("[Autostart] Installed {}", desktop_path.display());
-    Ok(())
-}
-
-fn uninstall_autostart_desktop() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    let desktop_path = autostart_desktop_path()?;
-    if !desktop_path.exists() {
-        return Err(format!("autostart entry not found: {}", desktop_path.display()).into());
-    }
-    std::fs::remove_file(&desktop_path)?;
-    println!("[Autostart] Removed {}", desktop_path.display());
-    Ok(())
 }
 
 /// Per-instance control dispatch mode. `Unicast` targets a single daemon bus
